+++ example/lib/widgets/asset_widget_builder.dart
@@ -3,7 +3,6 @@
 // in the LICENSE file.
 
 import 'package:flutter/material.dart';
-import 'package:photo_manager_image_provider/photo_manager_image_provider.dart';
 import 'package:wechat_camera_picker/wechat_camera_picker.dart';
 
 class AssetWidgetBuilder extends StatelessWidget {